Update VS Code Settings for Bolt (#8)
* Add heira-eyaml to VS Code recommended extensions

This commit adds an `extensions.json` file with the `brandontosch.hiera-eyaml` extension added. When a user opens this repo in VS Code, VS Code will notify them they should install this extension to use the functionality already configured in the `settings.json`.

* Update recommended extensions to include Puppet

This commit adds the `puppet.puppet-vscode` extension to the VS Code recommended extensions for this repo. If a user opens this repo with VS Code, it will prompt them to install the Puppet Extension, which will give them Bolt intellisense, hover information, and other features inside the Bolt plans. If a user does not have VS Code, this will not prompt them.

* Add Bolt config file validation

This commit adds the `redhat.vscode-yaml` extension to the recommendations and configures it in the workspace settings file. This will prompt a user to install the Redhat YAML extension when they open this project in VS Code, and point the extension to the Bolt YAML schema files. This adds intellisense, hover, and validation support for the Bolt configuration files as well as Bolt YAML plans.